Introduce multiple cursor functionality (#11)
Introduces several new heads, all of which are range-aware. The old f head (helm ag searches the current folder) has been renamed to d ('d' for "directory") to make room for the new head:

f: mark & next ('f' for "forward", but described as "next" for symmetry with the n head's name)
b: mark & prev ('b' for "backward", but described as "prev" for symmetry with the p head's name)
a: mark all. This replaces iedit, which is nice because
a) now we can remove that dependency (and replace it with something more standard)
b) it had no visual indication of cursors or the symbols that would be affected by edits
c) it was not range-aware (if I recall correctly)
e: edit marks. This is the head that the user should activate when they want to exit the hydra and start editing their file again, but with the multiple cursors active. C-g and q both exit the hydra but clean up all the cursors.
f, b, and a won't place redundant cursors -- this handles the situation where - for example - the user enters 'b', then 'b', and then wants to go forward, but hits 'f' instaed of 'n'.

The header now displays the number of multiple cursors active since the user's modeline my not display that info.

Regarding the save-excursion in symbol-navigation-hydra-mark-all... https://www.gnu.org/software/emacs/manual/html_node/eintr/save_002dexcursion.html doesn't say whether it handles exceptions but I included it in case it is nice and does, restoring the original point location in the exception-handling.

☑️ My elisp byte-compiles cleanly
☑️ M-x checkdoc is happy with my docstrings
✖️ I've used the latest version of package-lint to check for packaging issues, and addressed its feedback

I get the following errors for that last one:

```
7:23: error: Package auto-highlight-symbol is not installable.
7:54: error: Package hydra is not installable.
7:86: error: Package multiple-cursors is not installable.
```

This is referring to the Package-Requires line:

https://github.com/bgwines/symbol-navigation-hydra/pull/11/files#diff-f85f19de793c3e8152841193a7c46fc5R7

However, the latest MELPA build lists the dependencies (https://melpa.org/#/symbol-navigation-hydra) and they have links to those versions of the packages, so it was clearly happy with it as of 72982a7. When I run it on 72982a7, it fails with similar errors. Sooooooooo I think we're probably fine.

Fixes https://github.com/bgwines/symbol-navigation-hydra/issues/3

@riscy, if you get the chance sometime, I'd love to hear any feedback you have on the changes here! But no pressure/rush :)